+++ lib/data_factory/metadata_factory.rb
@@ -13,7 +13,8 @@
 #  limitations under the License.
 
 module OLE_QA::Framework
-  # This class contains methods to generate assorted metadata.
+  # This class contains methods to generate assorted metadata for the OLE application.
+  #   {OLE_QA::Framework::Bib_Factory} should be used for creating bibliographic metadata.
   class Metadata_Factory
     class << self
       include OLE_QA::Framework::Factory_Helpers